libminisip:  * Fix a few warnings in msvc
git-svn-id: svn://svn.minisip.org/minisip/trunk@3510 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/subsystem_media/rtp/RtcpReportRR.cxx
@@ -83,7 +83,7 @@ int RtcpReportRR::size(){
 
 
 int RtcpReportRR::get_n_report_blocks(){
-	return reception_blocks.size();
+	return (int)reception_blocks.size();
 }
                 
 RtcpReportReceptionBlock &RtcpReportRR::get_reception_block(int i){
