+++ libminisip/source/subsystem_media/rtp/SDES_CNAME.cxx
@@ -58,7 +58,7 @@ int SDES_CNAME::size(){
 //	int npad = 4-((2+cname.length())%4);
 //	if (npad==4)
 //		npad=0;
-	return 2+cname.length()/*+npad*/;
+	return 2+(int)cname.length()/*+npad*/;
 }
 
 #ifdef DEBUG_OUTPUT
